# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Incidents",
        "description": "API complète pour la gestion des incidents municipaux",
        "contact": {
            "name": "Support API",
            "email": "support@incidents.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/incidents/{incident_id}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Lister les commentaires d'un incident",
                "operationId": "b088c93a169f1b29e5cd5cccfb5476ea",
                "parameters": [
                    {
                        "name": "incident_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commentaires"
                    },
                    "404": {
                        "description": "Incident non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Créer un commentaire",
                "operationId": "da91d5f01dcdde15de6eae0126e11e8c",
                "parameters": [
                    {
                        "name": "incident_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Cet incident doit être résolu rapidement"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé"
                    },
                    "404": {
                        "description": "Incident non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/comments/{id}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Mettre à jour un commentaire",
                "operationId": "e6903ffe88ec53d1900c913ecab39ab0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commentaire mis à jour"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Commentaire non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "08965db07afd6799973a7023fc640b33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Commentaire non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum (format : Bearer <token>)"
            }
        }
    },
...
}